[refactor]: 라우팅 구조 변경(#3)
