+++ src/screen/instructor/RecordingDetail.tsx
